docs: refactor some enemy bundle stuff to new enemy module
